organize debug collector files
